Add package.json, .gitignore, and index.js
The package.json will allow us to share this module via [npm][npm].

The .gitignore keeps the dependencies from being committed. It depends
only on [React][react].

Lastly, index.js is the conventional entry point for accessing the
module. To allow keeping the source code in a subdirectory, we just
require the module from index.js.

[npm]: https://www.npmjs.com/
[react]: http://facebook.github.io/react/