Fall back to overflow style when finding scrollable parent
When taking a run at setting up Jest with @flarnie this afternoon, we
ran into an issue with the way that the Node environment reports the
computed style for 'overflow-y': it was undefined. 'overflow', however,
was defined. So, here we can fall back to overflow if overflow-y is
undefined, allowing this component to work in a wider variety of
environments. This is what we expect the browser to do anyway.