Add .projections.json file
This file defines alternates for JavaScript files. When used with
vim-projectionist, this enables easy switching between source and spec
files by running `:A`.

  https://github.com/tpope/vim-projectionist

One thing that tripped me up when I was working on this is it seems that
vim-projectionist and vim-rails both try to define `:A`, and vim-rails
always wins. So, in order to make this work, I needed to disable
vim-rails. There might be a better way to resolve this, but I didn't
have the time to figure it out properly. There's a little more
information at https://github.com/tpope/vim-rails/issues/326