Make react a peerDependency instead of a dependency
peerDependencies is appropriate for expressing the dependency between a
"plugin" and its "host package", to allow npm to resolve the dependency
tree in a way that makes sense for a single, top-level, "host package".
In this case, the "plugin" in question in react-waypoint, and the "host
package" is React.js.

More info: http://blog.nodejs.org/2013/02/07/peer-dependencies/