Configure `npm test` to run jshint and jscs
We'd like the code in this repo to maintain good style and quality, so I
am configuring npm test to run jshint and jscs. This should help common
problems get caught before code review.

I expect to expand this to include a Jasmine run as well, once we get
that set up.