Set up Karma, Jasmine, Webpack to run tests
When picking a testing setup, we first took a swing at setting up Jest.
This went pretty well until we ran into a problem with offsetParent and
offsetTop being undefined on all nodes. We believe this is because Jest
tests aren't run within an actual browser environment. We considered
mocking the browser behavior required for this component, but that
seemed too painful and brittle. Since this component is so tied to the
DOM, we decided that it made more sense to run our tests in an actual
browser.

So, I decided to give Karma a try. Karma is a framework-agnostic test
runner that can run tests in multiple browsers.

  http://karma-runner.github.io/

Since we are using CommonJS, we needed something that would resolve our
require statements. I first tried setting up karma-commonjs plugin since
it seemed pretty straightforward, however, I ran into a problem when
trying to require a dependency from node_modules. It turns out that this
plugin was designed to mimic basic Node CommonJS behavior, which does
not include looking in node_modules. More info:

- https://github.com/karma-runner/karma-commonjs/issues/29
- https://github.com/karma-runner/karma-commonjs/issues/37

So, I decided to try out karma-webpack and I was up and running in no
time. I borrowed most of the configuration from react-router. There are
some options in there for running in CI that we'll likely end up
tweaking when we get CI actually set up, but I anticipate that we will
be doing that soon so it seemed reasonable to start that work here.