Fix Waypoint in Firefox
The span that Waypoint renders to the page is rendered differently in
Firefox than Chrome. In Chrome, it has a height of 0, whereas in Firefox
it has a height that matches its font-size. This caused height
calculations in our specs to be off by 16px in Firefox. Giving this
element a font-size of 0 resolves this issue.