Fix window scrolling in Firefox
It turns out that my previous fix (a3dc459) for Waypoints with the
window element as their scrollable parent didn't fix the problem in all
browsers. In Chrome, it was all fine, because
`window.getComputedStyle(document)` returns null (and we added a guard
for that). In Firefox however, an error was thrown:

  TypeError: Argument 1 of Window.getComputedStyle does not implement
  interface Element.

I've changed the null-checking guard to an explicit check for if the
current node is the document element.

Specs are still failing for Firefox. I'm investigating that, but will
push this fix in the meantime.