Fix spec for window scroll
The spec that I previously wrote for when the scrollable parent is equal
to the window element (a3dc4593) only worked in Chrome. Our Travis
integration runs a Firefox instance, and in that browser the spec
failed. It turns out that Chrome will allow you to do `window.scrollTop
= x` and it will scroll to x. Firefox silently ignores that call. The
way we found to work in both browsers were `window.scroll(0, x)`.

Once that was fixed, I found another issue with our test setup. We
weren't cleaning up the divs we leave on the page. This caused the last
spec (testing scroll on the window element) to be contaminated by those
divs. We added an afterEach hook to unmount components as they were
tested.