Replace JSFiddle iframe with inline example on homepage
Instead of relying on JSFiddle, we can render an example in our own
page. This just moves the 'basic' example from JSFiddle into our own
repository, and makes the styles use BEM.

I had to inline the script at the bottom of the 'index.html' for the
in-browser JSX parser to parse it. Also removed module imports and
exports from the copy of 'waypoint.js' on this branch. Should have no
adverse affect on the cannonical version on master.

Future commits will add more example uses.