Add 'pointer-events: none' to message in basic example
In the "basic" example of React-Waypoint, we show a message overlay at
the top giving the status of the waypoint as you scroll the inner
window. Having your mouse on the message and scrolling should still
scroll the inner window, and adding 'pointer-events: none' to the
message is an easy fix that will work in most recent browsers.

While making this change, I sorted the CSS properties.