Use wrapper to auto-generate examples and code from inline scripts
We can easily grab the code for each demo component and
render it inline as a code sample below the demo component.

Props to @janpaul123 for this idea.

This commit also cleans up the actual JS of the scripts that generate
the demo components, since they will be rendered inline as code samples.