Automatically find component constructor for example demos
We have a wrapper that finds all examples and renders them with the
source code shown inline below the example, automatically.

To make it work even better, so that adding examples is even easier in
the future, this updates it to automatically find the constructor of a
given component in the global namespace, rather than relying on a series
of conditionals which would need updated for new examples.