Use BookmarkableHeader component in examples; fixes #6
We want our examples to be bookmarkable, so this component adds
bookmarkable headers in the style of Wikipedia.

It would be nice if all our headers in the documentation were
bookmarkable - resusing the BookmarkableHeader component there would be
easy if the whole content of the page were rendered by React.

We may be re-organizing the project to use the build system, and if we
do that on the github pages branch then we can more easily split the
inline scripts into separate files. Rendering the whole page content as
a React component will follow naturally from that step. (See issue #7)