Listen to 'resize' on window, not scrollableParent; fixes #8
(From the relevant issue:)
We listen to the 'scroll' and 'resize' events on the 'scrollableParent'
element, which is sometimes the window. However, the 'resize' event is
[only triggered on the window element][mdn-resize-event], so we may not
catch it in cases that the 'scrollableParent' is not the window.

It seems safer to listen to the 'resize' event always on the window,
rather than the scrollable parent.

[mdn-resize-event]: https://developer.mozilla.org/en-US/docs/Web/API/GlobalEventHandlers/onresize