Fire callbacks when scrolling really fast
When scrolling really fast, there was a chance the onEnter and onLeave
callbacks weren't fired, despite the Waypoint being "visible" (in
viewport) for a brief moment.

To test, you can do this in your browser:

  window.addEventListener('scroll', function() { console.log(window.scrollY) });

Then try scrolling really fast. You could see something like this in
your console log:

  2
  648
  939
  3775

With the previous implementation and a Waypoint located at 2000, no
callbacks would (probably, depending on screen size and threshold
setting) have been fired in this scenario.

To fix this, I've switched to a model where we keep track of the
"position" of the Waypoint. Either `above`, `below` or `inside`. If we
go directly from `below` to `above` (and vice versa), we fire both
callbacks in sequence.

Since separating programmatic scrolls (e.g. where you call
`window.scroll`) from user-initiated scrolls is hard [1], I haven't done
anything to deal with differentiating them. This has the effect that
even programmatically generated scrolls, e.g. jumping to a section, will
trigger onEnter and onLeave for all Waypoints along the way. This is
potentially a serious issue, but I can't say for sure.

Addresses #25.

[1]: http://stackoverflow.com/questions/542973/how-to-i-detect-if-an-onscroll-event-was-programmatically-generated