Add test to verify that onEnter is called only once
We've had reports of onEnter firing multiple times while in view
(e.g. #28). I've been trying to reproduce the bug with a test, but I
can't reproduce it. I'm adding the test that I wrote anyway as it will
prevent possible regressions. It might also inspire someone else to dig
in and find other things that I didn't find.