Document some limitations in the readme
At Brigade, we recently bumped into an issue where <Waypoint> would
always call `onEnter` no matter if it was visible or not. We determined
that this was due to a change we made that caused its scrollable
container to have a height that was both larger than the window's
viewport height and so that the container did not actually scroll (it's
overflow was still set, but it's height was tall enough for whatever
content it received). We brainstormed some ideas for how to resolve this
in react-waypoint, but the we are worried about the performance
implications of the solutions. We believe that this may be enough of an
edge case that we don't actually need to solve right now, so we decided
to document this in the readme to help others who might be running into
this issue. If this becomes more of a problem for more people, we might
consider implementing a better solution in code.