Prevent duplicate callback calls
If your onEnter/onLeave callback made the <Waypoint> re-render, you
could end up with duplicate callback calls. This was because we were
remembering the previous position too late. Consider the following
example:

  <List>
    {items.map((item) => <Item item={item}/>)}
    <Waypoint onEnter={() => this.setState({ foo: 'bar' })}/>
  </List>

When the waypoint is in view, it will fire the onEnter callback. The
setState call will then make the waypoint update/re-render. Since we are
checking waypoint position inside componentDidUpdate, we call onEnter
again.

I thought about adding one or a few defers here and there to fix this
problem (setTimeout(this._handleScroll)), but I ran into a few problems
with getting all specs to pass. Also, I'm a little bit afraid to break
some implementations that rely on not being deferred. So I took the easy
way out and just moved the remembering of previous position higher up so
that the second call would correctly notice the unchanged of position.

Thanks to @imouto1994 for providing enough context so that I could
create a repro test case. Also thanks to @chengyin who reported a
similar issue (#28) and provided enough details there too. But at that
time I wasn't able to create a repro case.

Fixes #35