Ensure component is mounted when executing scroll handler
In the Brigade app, we've seen an error occasionally happen in
production where `this === null` inside of the `_currentPosition`
function. This causes `ReactDOM.findDOMNode` to error out.

While I'm not entirely sure I understand how this could happen, my
hypothesis is that the component has become unmounted before the scroll
handler can be executed, causing a race condition. Ensuring that the
component is mounted before executing the handler should prevent this
from happening.