Upgrade Babel 5 -> 6
The new version of Babel is much more modular, requiring a few extra
packages to be installed to give us the functionality we had before with
whatever used to be built into the core Babel package. I decided to just
roll with the ES2015 preset instead of picking just the packages that we
actually are using, since ES2015 is a general expectation that we have
and I don't want to make future developers confused.