Convert to ES2015 class
ES2015 classes have fewer features than React.createClass (e.g. no
mixins and no anti-pattern this.isMounted), which I see as an advantage
of their use. I've also read that they are slightly more performant,
which might be a nice side-benefit. Since it is simple enough to switch
here, it seemed like a good thing to do.

In this conversion, I had to remove the isMounted guard that I added in
4bbdfa430 since it is not available in ES2015 classes. However, this
method is an anti-pattern and on its way to being deprecated anyway so
it should be generally avoided [0].

Also, since auto-binding isn't a feature of classes, I added a binding
of the scroll handler to the constructor, as is suggested by the React
0.13.0 beta 1 release notes [1].

[0]: https://facebook.github.io/react/blog/2015/12/16/ismounted-antipattern.html
[1]: https://facebook.github.io/react/blog/2015/01/27/react-v0.13.0-beta-1.html#autobinding