Make bundled waypoint.js easier to import
With Babel 6, the default export behavior changed [1] so that there's no
longer a `module.exports = ...`. This causes problems down the line,
because other projects that depend on react-waypoint will need to import
it through

  const Waypoint = reauire('react-waypoint').default;

(notice the added `.default`)

Thankfully, there's a plugin that will automatically inject the
`module.exports` for us [2]. I've added and configured that plugin in
this commit.

[1]: https://phabricator.babeljs.io/T2212
[2]: https://www.npmjs.com/package/babel-plugin-add-module-exports