Throw error if using legacy `scrollableParent` prop
I just renamed the `scrollableParent` prop to `scrollableAncestor`.
Since using a non-existing prop in React won't result in any warning of
any kind, there was a risk that people would upgrade and then end up
with hard-to-trace bugs. To hopefully prevent a few of those cases, I'm
adding code that throws an error when using the old prop.

Thanks to @janpaul123 for chiming in to the conversation and convincing
me this was the right thing to do.