Disable eslint proptype check for `scrollableParent`
I was running into an eslint error because of the added check for the
discontinued prop `scrollableParent`. I don't want this to be part of
the propTypes declaration so in order to avoid the error I'm adding a
comment to disable it locally.

While I was at it, I fixed another eslint warning in the same file (line
too long).