Handle waypoint parent not being visible
This can happen for a variety of reasons, e.g. a transition animation.

Rather than throwing an exception,m we set the position to the new "invisible"
position. When the component becomes visible and we notice in the next scroll
event (or forced update), an standard event handlers will fire as we notice
the change from "invisible" to whatever the new state is.