Unify arguments passed to callbacks
This is a breaking change. Instead of requiring consumers of the
Waypoint interface to remember the order of the arguments passed to
onEnter/onLeave/onPositionChange, I'm combining everything into
an object.

Before this change, you would see things like

   <Waypoint
     onEnter={(event, from) => { ... }}
     onPositionChange={(to, from, event) => { ... }}
   />

After this change, you'll do this instead:

   <Waypoint
     onEnter={({ previousPosition, event }) => { ... }}
     onPositionChange={(currentPosition, event) => { ... }}
   />

Or if you are not using object deconstructing:

   <Waypoint
     onEnter={(properties) => { ... properties.previousPosition }}
     onPositionChange={(properties) => { ... properties.event }}
   />

This is nice for a few reasons:
 - You no longer have to "ignore" the first argument (event) for
   onEnter/onLeave.
 - All callbacks fulfill the same contract.
 - It will be easier going forward to add more information to the
   argument.

Even though it's a little bit redundant to pass `currentPosition` to
`onEnter`, and `previousPosition` to `onLeave`, I decided to keep them
there to make everything consistent.