Add Troubleshooting info to README.md
@bySabi made a good point about the `debug` prop polluting the code.
https://github.com/brigade/react-waypoint/issues/93#issue-158499180

In general, I agree. But in this case I think the alternative (clone and
modify the source code) isn't always that straightforward. However, we
can help people who want to take that approach by adding a section to
the readme. I basically copy-pasted what @bySabi had already listed, and
added a `babel` step.