Fix test for `throttleHandler`
The previous test was partly failing because the component was unmounted
before we could do the assertions. I've fixed that by adding an
`afterEach` block with a delay (slightly longer than the execution
time). We could probably solve this with jasmine.clock(), but I think
using a setTimeout is more straightforward. And we're not introducing
any noticeable delays.