Avoid React warning about propTypes
In the latest version of React, <Waypoint> is causing a lot of these
warnings to show up in the console log:

application.self-8f8100c….js?body=1:1512
Warning: You are manually calling a React.PropTypes validation function
for the `debug` prop on `Waypoint`. This is deprecated and will not work
in the next major version. You may be seeing this warning due to a
third-party PropTypes library. See
https://fb.me/react-warning-dont-call-proptypes for details.

I think we can avoid this by not assigning propTypes to a variable
before setting them on the React class.