Manage a single window resize event instead of one per component
I am looking into improving the performance of react-waypoint. We can be
a little more mindful of how many event listeners we end up creating by
only managing a single event listener. Using Chrome's timeline tool, I
have verified that this type of optimization is significant when there
are a lot of event listeners on the page.

Since we always set up a window resize event, this seemed like a good
place to start for this optimization. I originally chose a Set, since
that seemed like the easy and nice way to go, but it will require that
older browsers have a polyfill so I decided to take a different
approach. By maintaining an object with an index, and using the index to
remove the event handler from the object, we can get similar performance
as the Set, but with broader compatibility.

If we like this approach, I plan to do something similar for the scroll
event listeners. These are a little more complicated since they could be
attached to any scrollable container, instead of just the window. One
option would be to use a Map to manage a Set of event listeners for each
scrollable container. It might also be worth exploring using a single
top-level event listener instead of one per scrollable container, which
should be accomplishable via the Map/Set approach.