Move some "private" methods off the Waypoint class
These methods don't reference `this`, so it is easy enough to move these
off the class. This should make the code a bit clearer and improve
performance a little.

There are probably a few more methods that could be migrated this way
with a little more effort, but I decided to not do that at this time.