Improve publishing experience
In preparation for publishing v4, I wanted to bring over some of the npm
scripts I had set up for react-with-styles. This should make publishing
the next version a bit smoother.

I also needed to make Karma run as a single run by default, so I added a
testonly:watch script that allows you to run it in non-single-run mode
if you like.