Avoid leaking event listeners
In c6711294 I made a change to waypoint that makes it so it sets up
fewer event listeners. Unfortunately, I did this in a way that causes
the added event listener to never be removed, as @trotzig pointed out at

  https://github.com/brigade/react-waypoint/commit/c6711294#commitcomment-19498628

This is happening because every time we call .bind() on a function, we
get a new object, which the browser is unable to use to remove the
previous function. The solution is to bind the function once, and keep
track of it so we can reuse the reference when we want to remove the
event listener.