Flatten some test cases
I think I found a bug so I am rolling through the test cases to see what
we have covered and I'm finding it a little difficult to see what we are
testing due to the increased indirection caused by nested describe
blocks. To make this a little easier for future explorers, I'm
flattening out some of these examples. There's more flattening that
should be done here, but this seems like a good step in the right
direction.