Remove support for throttleHandler prop
This prop has had people confused. It was added to allow people to
better control scrolling performance in their applications. Since we
added it, we've done other changes to improve performance (3ad940e,
c67112) which has made the benefit of throttling questionable. Removing
it will reduce the complexity of the component.