Remove some unnecessary indirection in tests
I was thinking about making a change to waypoint, but I found the tests
to be unnecessarily difficult to follow, so I decided to first reduce
some of the indirection and nesting. I think there is more that could be
done here, but this seems like a pretty good start.

I am trading indirection for a little duplication--for tests this is
usually the right tradeoff to make.