Avoid unnecessary clearTimeout when unmounting
I was doing some profiling and noticed a pattern on pages that have a
lot of waypoints. In the timeline I can see that we spend a non-trivial
amount of time in clearTimeout (~0.25ms per waypoint on my laptop). This
can add up, so let's avoid the extra work by adding a check.