Drop support for React v14
With React 16 soon to be released, I'm going through a process to
prepare the codebase for that version. Keeping support for React 14
isn't really hurting us at this moment, but it's going to be more
painful going forward.

Projects that are still on React 14 will have to lock their version of
react-waypoint to something <= 5.2.1.