Make clear that webpack setup is for perf test
People coming across a webpack.config.js file in the root of this repo
might assume that we use webpack to build waypoint. Since we only use it
to create the bundle for the performance test page, we can rename the
config file to make things clearer.

I also added a comment at the top of the file to explain why it was
there.