Explain why .importjs.js is there
I'm so blinded by always adding this config file to all my projects, so
it never crossed my mind to explain why it was there. Adding a comment
at the top with a link to the import-js github repo should make things
clearer.

Thanks to @jmeas for pointing this out in code review.