Mount and unmount components in perf-test
I'm trying to force bad performance out of a page in order to highlight
performance problems that we could solve in the current implementation.
This commit adds state to the wrapping component, conditionally
rendering a component with an inline svg or a placeholder div.

After this change, I'm noticing waypoint's _getBounds method take 3-4%
of time while scrolling, all coming from componentDidUpdate. I'm
starting to think that we might be able to add a prop or something to
disable calculating bounds on update. Often times, the waypoint itself
might not have moved on an update so there's no need checking if we need
to trigger anything.