Assume `window` as scrollableAncestor for <body>
We've seen reports (e.g. [1], [2]) where people have to add
`scrollableAncestor={window}`, because their <body> element has an
overflow style that is either "auto" or "scroll". Unless I'm mistaken,
there's no way to make the body scroll independently, which means that
we can assume that the element to attach scroll listeners to is
`window`.

Making this assumption led to a bit of simplification around a few
exceptions we had made when traversing up the DOM tree. Once we reach
`document.body`, we simply return. I can't think of a scenario where we
would want to continue traversing from that node.

[1] #171
[2] #182