Improve startup time
While doing some profiling, I noticed that on pages that render a lot of
waypoints, a lot of time is spent in componentDidMount. This is mostly
due to the cost of calling setTimeout, but also because of the other
work we are doing in here, like findScrollableAncestor.

By consolidating the setTimout calls, we greatly reduce the cost of
componentDidMount. We can take this a step further and defer
initializing the event listeners and finding the scrollable ancestors
since we won't need any of that anyway until our initial handleScroll
call can happen. This effectively makes componentDidMount free, which
frees up the main thread for more important work, like rendering pixels.

I believe that this also allows the calls to findScrollableAncestor to
work faster by avoiding some layout thrashing. Since this function
traverses up the DOM tree and computes styles on a bunch of nodes, this
is important. When these operations happened in componentDidMount, it is
possible that other components would invalidate layout in between these
calls from multiple waypoints. By clustering all of them together, we
are less likely to have to recalculate style or compute layout every
time.

Additionally, there may be cases where waypoints are rendered and
unmounted very quickly. In this case, we might actually be able to avoid
some of the unnecessary work we are doing here, which will be a small
win for those situations.

The main drawback to this is that if you have a lot of waypoints on a
page, the previous approach keeps each initial run a bit more async,
which allows the browser to do work in between. Now, all of that work is
in the same block. I think this is an acceptable tradeoff.