Build with rollup
Previously, we were building waypoint just by running it through Babel.
Since we split things out into separate modules, this caused a bit of
extra overhead in the bundle size and at runtime due to the extra
wrapping of these modules.

Thankfully, rollup was built to help us produce a smaller build. To make
this work, I needed to remove the Babel configuration that specifies a
module transformer, so I decided to switch to the Airbnb Babel preset,
which exposes an option for this (and under the hood uses
babel-preset-env, which we need to migrate to anyway). More info:

  https://github.com/rollup/rollup-plugin-babel#configuring-babel

The rest of this was loosely based off the rollup-starter-lib
configuration. This gives us an ESM build pretty much for free.

  https://github.com/rollup/rollup-starter-lib/tree/babel