Add message to better understand log message in tests
We have a test that verifies that an error is thrown as part of
rendering a waypoint with more than one child. Even though we take
measures to prevent the error from leaking outside the test itself,
Karma picks it up and prints a confusing log:

ERROR: 'The above error occurred in the <Waypoint> component:
    in Waypoint
    in div

No error is logged above that line, which makes it hard to understand.

While it's probably possible to get rid of that line some other way, I
opted for a simple solution to begin with, simply logging something that
makes it A) easier to ignore B) easier to find the offending test case.