Stop running tests before releases
This step is causing issues for me right now. It's a little tricky to
set up the test suite locally, and requiring those to pass before a
release is made is a little aggressive. Tests are usually run in CI
anyway, so this step is mostly just in the way.